Early Player
Implemented a Player script that contains an abstract class for properties, like walking speed. Added an abstract class for data, like current health or velocity. Created an abstract class for the playerController, responsible for the functions of movement, for example. And implemented the Player script that receives input from the player directly and store interaction information, like rays from the camera (This script is equivalent to a PlayerInput script mostly). Created a tag for interactable objects.