Llama GetCaptured
Implemented function in Llama to be captured. Upon capture the Llama will trigger a custom Event type that will send its id to the function invoked. The listener to this event is add by the LlamaManager, that way the Llama script does not depend on anything to work, but it has to be setup (which is okay because that already happens during pooling).